refactor machine to /machines/commerce
